+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.12.2] - 2026-6-25
+
+### Improvements
+
+ - 🔧 Fixed bug where the slider frame range for selected sensor is based on max of frames of detections and tracks across all sensors.
+ - 🛠️ Removed `sensors` as a programmatic argument. It is already required for the `Imagery`, `Detector` and `Track` objects. The unique sensors list is determined from these.
+
 ## [1.12.1] - 2026-5-22
 
 ### Improvements
@@ -359,6 +366,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed bug where tooltips did not take into account imagery row / column offsets.
 - Fixed bug where imagery produced by algorithms did not have pre-computed histograms (which improves playback performance)
 
+[1.12.2]: https://github.com/awetomaton/VISTA/releases/tag/1.12.2
+[1.12.1]: https://github.com/awetomaton/VISTA/releases/tag/1.12.1
 [1.12.0]: https://github.com/awetomaton/VISTA/releases/tag/1.12.0
 [1.11.1]: https://github.com/awetomaton/VISTA/releases/tag/1.11.1
 [1.11.0]: https://github.com/awetomaton/VISTA/releases/tag/1.11.0